Add user and group options
If you are running cmdline as uid 0, by default every command will run
as user "nobody" for security purposes. Unless you specify user, which will
be used to execute the command.

If option group is not specified, it takes the user's main group. Otherwise
it will make sure the user is within the group.

User or uid can be used as well as group or gid.