Add max-running configuration
Max-running enables you to set the maximum number of command to execute
at the same time.

It could be useful if you have 5 CPU intensive commands to run, when you start
cmdlinebeat it will run those 5 commands at the same time. So max-running will
avoid that.

By default it is 0 which is run everything as fast as you can